pyproject.toml
[project]
name = "fyp25-literature-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "hh65", email = "hh65@sussex.ac.uk" }
]
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.85",
    "dotenv>=0.9.9",
    "loguru>=0.7.3",
    "openai>=2.5.0",
    "openai-agents>=0.4.0",
    "rich>=14.2.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=7.0.1",
    "jupyter>=1.1.1",
    "loguru>=0.7.3",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.14.1",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
